feat : 자음으로 연관검색 기능구현

+++ controllers/posts.controller.js
@@ -1,6 +1,7 @@
 const PostService = require("../services/posts.service");
 const { Hashtag } = require("../models");
 const log = require("../config/logger");
+const help = require("korean-regexp");
 let BadRequestError = require("./http-errors").BadRequestError;
 class PostController {
   postService = new PostService();
@@ -27,15 +28,18 @@ class PostController {
         hashtag,
         userId
       );
-
-      let arr_hashtag = hashtag.toString().split(",");
-      for (let i = 0; i < arr_hashtag.length; i++) {
-        const createHashtag = await Hashtag.create({
-          hashtag: arr_hashtag[i],
-          postId: createPost.postId,
-        });
+      let consonant = [];
+      if (hashtag) {
+        //유저가 Hashtag를 입력했을 때 Hashtag table에 Hashtag를 생성
+        for (let i = 0; i < hashtag.length; i++) {
+          consonant[i] = help.explode(hashtag[i]).join("");
+          const createHashtag = await Hashtag.create({
+            hashtag: hashtag[i],
+            consonant: consonant[i],
+            postId: createPost.postId,
+          });
+        }
       }
-
       res.status(201).json(createPost);
     } catch (error) {
       next(error);
@@ -45,7 +49,12 @@ class PostController {
     try {
       const { pagenum } = req.params;
       const { userId } = req.body;
-
+      if (!pagenum) {
+        log.error("PostController.getAllPosts : pagenum is required");
+        throw new BadRequestError(
+          "PostController.getAllPosts : pagenum is required"
+        );
+      }
       let type = false;
       const getAllPosts = await this.postService.getAllPosts(pagenum, userId);
 
@@ -84,12 +93,20 @@ class PostController {
     try {
       const { postId } = req.params;
       const { content, time, distance, path, speed, image, hashtag } = req.body;
+      if (!postId) {
+        log.error("PostController.updatePost : postId is required");
+        throw new BadRequestError(
+          "PostController.updatePost : update is required"
+        );
+      }
       let arrayHash = [];
       let checkHash = false;
       const checkSameHashTag = await Hashtag.findAll({
         where: { postId },
         attributes: ["hashtag"],
       });
+
+      //유저가 게시글을 수정할 때 hashtag를 수정하면 Hashtag 테이블의 hashtag도 같이 수정
       for (let i = 0; i < checkSameHashTag.length; i++) {
         arrayHash.push(checkSameHashTag[i].hashtag);
       }
@@ -123,6 +140,12 @@ class PostController {
   deletePost = async (req, res, next) => {
     try {
       const { postId } = req.params;
+      if (!postId) {
+        log.error("PostController.deletePost : postId is required");
+        throw new BadRequestError(
+          "PostController.deletePost : update is required"
+        );
+      }
       const deletePost = await this.postService.deletePost(postId);
 
       if (deletePost === 0) {
@@ -138,6 +161,12 @@ class PostController {
     try {
       const { pagenum } = req.params;
       const { userId } = req.body;
+      if (!pagenum) {
+        log.error("PostController.getLikeAllPosts : pagenum is required");
+        throw new BadRequestError(
+          "PostController.getLikeAllPosts : pagenum is required"
+        );
+      }
       let type = false;
       const getLikeAllPosts = await this.postService.geLikeAllPosts(
         pagenum,
